Move createPunishmentTableRow
